requirements.txt
flask>=2.3.0
python-chess>=1.9.0
torch>=2.0.0
numpy>=1.24.0
stockfish>=3.28.0
